# crates/librtbit/webui/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "rtbit API",
    "description": "REST API for rtbit BitTorrent client",
    "version": "7.0.0"
  },
  "paths": {
...
    "/torrents/limits": {
      "get": {
        "summary": "Get rate limits",
        "description": "Get current upload/download rate limits.",
        "tags": ["Torrents"],
        "responses": {
          "200": {
            "description": "Current limits",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LimitsConfig" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Set rate limits",
        "description": "Update upload/download rate limits.",
        "tags": ["Torrents"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LimitsConfig" }
            }
          }
        },
        "responses": {
          "200": { "description": "Limits updated" }
        }
      }
    },
...
    "/torrents/resolve_magnet": {
      "post": {
        "summary": "Resolve magnet",
        "description": "Resolve a magnet link to torrent metadata bytes.",
        "tags": ["Torrents"],
        "requestBody": {
          "content": {
            "text/plain": {
              "schema": { "type": "string" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Torrent file bytes",
            "content": {
              "application/octet-stream": {
                "schema": { "type": "string", "format": "binary" }
              }
            }
          }
        }
      }
    },
...
    "/torrents/{id}/playlist": {
      "get": {
        "summary": "Torrent playlist",
        "description": "Generate M3U8 playlist for streamable files in this torrent.",
        "tags": ["Torrent"],
        "parameters": [
          { "$ref": "#/components/parameters/torrentId" }
        ],
        "responses": {
          "200": {
            "description": "M3U8 playlist",
            "content": {
              "application/x-mpegurl": {
                "schema": { "type": "string" }
              }
            }
          }
        }
      }
    },
    "/torrents/{id}/stream/{file_id}": {
      "get": {
        "summary": "Stream file",
        "description": "Stream a file from the torrent. Supports Range header for seeking.",
        "tags": ["Torrent"],
        "parameters": [
          { "$ref": "#/components/parameters/torrentId" },
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "File content",
            "content": {
              "application/octet-stream": {
                "schema": { "type": "string", "format": "binary" }
              }
            }
          },
          "206": {
            "description": "Partial content (range request)"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "torrentId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Torrent ID (integer) or info hash (hex string)",
        "schema": { "type": "string" }
      }
    },
    "schemas": {
...
      "LimitsConfig": {
        "type": "object",
        "properties": {
          "upload_bps": { "type": "integer", "nullable": true },
          "download_bps": { "type": "integer", "nullable": true }
        }
      }
    }
  }
}